requirement.txt
asgiref==3.3.4
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
decorator==5.0.9
defusedxml==0.7.1
Django==3.2.4
django-rest-knox==4.1.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
djongo==1.3.6

dynaconf==3.1.4
hvac==0.10.14
idna==2.10
itypes==1.2.0
Jinja2==3.0.1
loguru==0.5.3
MarkupSafe==2.0.1
oauthlib==3.1.1
pycparser==2.20
PyJWT==2.1.0
pymongo==3.11.4
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
shortuuid==1.0.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.2.4
uritemplate==3.0.1
urllib3==1.26.5
validators==0.18.2
win32-setctime==1.0.3
